% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_latex.R
\name{read_latex}
\alias{read_latex}
\title{Read latex output into R}
\usage{
read_latex(latex_output, output = "coef")
}
\arguments{
\item{latex_output}{The latex code to be converted into R objects}

\item{output}{A character vector specifying which type of regression output 
should be converted to an R object. Options are 'coef', 'stats', and 'extra'}
}
\description{
This function reads in Latex code of a regression summary into R
}
\examples{
height <- runif(100, 60, 78)
dad_height <- runif(100, 66, 78)
mom_height <- runif(100, 60, 72)

model <- lm(height ~ dad_height + mom_height)
latex_output <- regtable(list(model), 
   est = c('dad_height', 'mom_height'), 
   output_format = "latex")

latex_coef <- read_latex(latex_output, output = 'coef')

}
