% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtable_stack.R
\name{regtable_stack}
\alias{regtable_stack}
\title{Stack Multiple Regressions Outputs}
\usage{
regtable_stack(final_tables, table_names = NULL,
  output_format = "latex", note = NULL, header = NULL)
}
\arguments{
\item{final_tables}{list of output produced by regtable, which includes 
\itemize{
  \item \emph{output}: a dataframe version of the regtable results
  \item \emph{model_names}: either the character vector specified by 
    \emph{mnames} in regtable or a function generated vector of numeric 
    IDs for the models
} 
To output a dataframe in regtable, pass in "df" for the output_format.}

\item{table_names}{A character vector specifying the names of the objects in
final_tables. These names are concatenated with the coefficent names in 
the output.}

\item{output_format}{A string passed to kable() that specifies the format of
the table output. The options are latex, html, markdown, pandoc, and rst.
The default is latex.}

\item{note}{A character string if a footnote is to be added to the end of the
table.}

\item{header}{A character vector to be passed into 
\code{\link[kableExtra:add_header_above]{add_header_above}} that creates
a new header row. This should have length equal to ms from regtable()}
}
\description{
This routine takes in multiple outputs from regtable and 
arranges them into one output. The function stacks the different coefficients 
and summary statistics, while the extra_rows are grouped at the bottom of the 
regtable_stack() output.
}
\examples{
# create covariates
x1 <- rnorm(1000)
x2 <- rnorm(length(x1))

## fixed effects
fe <- factor(sample(20, length(x1), replace=TRUE))

## effects for fe
fe_effs <- rnorm(nlevels(fe))

## creating left hand sides y1 and y2
u <- rnorm(length(x1))
y1 <- 2 * x1 + x2 + fe_effs[fe] + u
y2 <- 3 * x1 + x2 + fe_effs[fe] + u

m1 <- felm(y1 ~ x1 + x2 | fe)
m2 <- lm(y1 ~ x1 + x2)

n1 <- felm(y2 ~ x1 + x2 | fe)
n2 <- lm(y2 ~ x1 + x2)

## generate output from regtable
r1 <- regtable(list(m1, m2), est = "x1", 
       output_format = "df")
r2 <- regtable(list(n1, n2), est = "x1", 
       output_format = "df")

regtable_stack(list(r1, r2), table_names = c("1", "2"), output_format = "rst")


}
