% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest_utils.R
\name{rf_semipar}
\alias{rf_semipar}
\title{Random Forest Helper Functions}
\usage{
rf_semipar(formula, data, ...)
}
\arguments{
\item{formula}{The regression formula to evaluate}

\item{data}{The dataframe that contains all the variables used in the formula}
}
\value{
An estimated model.
}
\description{
Estimate a partially linear model using the Robinson 1988 methodology using
random forests as the nonparametric regression technology.  The formula
structure has three parts: y ~ x | z.  y is the outcome, x are the linear
variables and z are the non-parametric variables.  Tt *should* work with
whatever transformations you throw at it, but note that the resulting lm
object it returns may have column names that are slightly different if you do
transformations.  For example, the variables in the output of a model
drilled ~ Auction + log(acres) | Cent_Lat + Cent_Long + EffDate will be
Auction and *log_acres_*.  ALSO: if you put something more complicated in,
like, say, a call to bs() from the splines package, you are going to get
pretty crazy names, but it should still work.
}
\examples{
library(tibble)
# create mock data
height <- runif(100, 60, 78)
dad_height <- runif(100, 66, 78)
mom_height <- runif(100, 60, 72)

df <- tibble(height = height, dad_height = dad_height, 
   mom_height = mom_height)

# evaluate model using rf_semipar
model <- rf_semipar(height ~ dad_height | mom_height, data = df)
}
