% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_latex.R
\name{split_latex}
\alias{split_latex}
\title{Read Sections of Latex Output into R}
\usage{
split_latex(latex, output = "coef")
}
\arguments{
\item{latex}{The latex code to be converted into R objects}

\item{output}{A character vector specifying which type of regression output 
should be converted to an R object. Options are "coef", "stats", and "extra". 
"coef" refers to regression coefficients and standard errors, and "stats" 
refers to other statistics such as N and R^2. "extra" should be used if the 
latex code you want to read into R was created by regtable() using the 
extra_rows argument, and you're interested in seeing the content of 
those extra rows.}
}
\value{
A dataframe with specific regression information.
}
\description{
This function reads in latex code of a regression summary into 
R, returning a subset of the latex code representing regression coefficients, 
statistics, or extra information.
}
\examples{
library(lfe)

# create covariates
x1 <- rnorm(1000)
x2 <- rnorm(length(x1))

# fixed effects
fe <- factor(sample(20, length(x1), replace=TRUE))

# effects for fe
fe_effs <- rnorm(nlevels(fe))

# creating left hand side y
u <- rnorm(length(x1))
y <- 2 * x1 + x2 + fe_effs[fe] + u

# evaluate models
m1 <- felm(y ~ x1 + x2 | fe)
m2 <- glm(y ~ x1 + x2)

# format models using regtable
latex <- 
   regtable(list(m1, m2), est = list("x1", c("x1", "x2")), 
   stats = list(c("adj.r.squared"), c("AIC")),
   stats_names = list(c("$Adj R^2$"), c("AIC")), 
   sig_stars = TRUE, output_format = "latex")

# read in regtable coefficients
latex_coef <- split_latex(latex, output = 'coef')

# read in regtable statistics
latex_stats <- split_latex(latex, output = 'stats')

}
