% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtable.R
\name{regtable}
\alias{regtable}
\alias{regtable_stack}
\title{Format R regression output}
\usage{
regtable(ms, est, mnames = NULL, est_names = NULL, extra_rows = NULL,
  se_fun = vcov, stats = c("r.squared", "adj.r.squared"),
  stats_names = c("$R^2$", "Proj. $R^2$"), n_obs = TRUE,
  output_format = "latex", header = NULL, sig_stars = FALSE,
  decimals = NULL, note = NULL, ...)

regtable_stack(final_tables, table_names = NULL,
  output_format = "latex", note = NULL, header = NULL)
}
\arguments{
\item{ms}{one or more model objects that are compatible with the summary()
function}

\item{est}{a character vector of the covariates to output}

\item{mnames}{a character vector of labels for each model object in ms}

\item{est_names}{a character vector of labels for est}

\item{extra_rows}{a character vector additional information to display for
each model, such as fixed effects or controls.}

\item{stats}{a character vector specifying which model statistics should be
kept in the output}

\item{stats_names}{a character vector of labels for each object in stats}

\item{output_format}{A string passed to kable() that specifies the format of
the table output. The options are latex, html, markdown, pandoc, and rst.
The default is latex}

\item{sig_stars}{logical indicating whether or not significant stars should
be added to the coefficients}

\item{note}{A character string if a footnote is to be added to the end of the
table.}
}
\description{
This routine creates Latex code, HTML code, and text tables for
output of a regression summary
}
\examples{
require(stats)
library(tidyverse)
model <- mtcars \%>\% lm(disp ~ drat + cyl, data = .)
latex_output <- regtable(list(model), 
   est = c('drat', 'cyl'), 
   output_format = "latex")

}
